test(schema): relocate predeclared-attr regression to trails sibling
Move the trails-only regression (no Rails counterpart) out of the
Rails-mirrored persistence.test.ts into persistence.trails.test.ts, and
make it a reliable fail-without-fix guard: evict `aircraft` from the
adapter schema cache so the reconcile path deterministically cannot find
its columns (a warm cross-file cache otherwise masks the dropped write).
Assert the foreign `expires_at` column no longer leaks into the subclass
and the predeclared virtual attribute survives reflection.

Also tighten the ensureSchemaLoaded comment to describe both foreign-map
detection paths (inherited owner vs forked copy).
